% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vuln_map_function}
\alias{vuln_map_function}
\title{Function to add Polygons of neighborhood vulnerability onto a base map}
\usage{
vuln_map_function(base_map_input, vuln_data, pal_input)
}
\arguments{
\item{base_map_input}{The base map}

\item{vuln_data}{The dataset to be used for the polygons
This dataset should have columns 'vulnerability_quantiles' & 'lsoa11_name'}

\item{pal_input}{The pallette to be used for the polygons}
}
\description{
Function to add Polygons of neighborhood vulnerability onto a base map
}
